build: move header files around a bit more, slightly rationalizing them
