+++ include/pocketsphinx.h
@@ -49,10 +49,10 @@
 #include <sphinxbase/feat.h>
 
 /* PocketSphinx headers (not many of them!) */
-#include <pocketsphinx_export.h>
-#include <cmdln_macro.h>
-#include <ps_lattice.h>
-#include <ps_mllr.h>
+#include <pocketsphinx/export.h>
+#include <pocketsphinx/cmdln_macro.h>
+#include <pocketsphinx/ps_lattice.h>
+#include <pocketsphinx/ps_mllr.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -66,7 +66,7 @@ extern "C" {
  */
 typedef struct ps_decoder_s ps_decoder_t;
 
-#include <ps_search.h>
+#include <pocketsphinx/ps_search.h>
 
 /**
  * PocketSphinx N-best hypothesis iterator object.
