+++ include/pocketsphinx/export.h
