+++ include/pocketsphinx/ps_search.h
